With a little less stupidity, this is looking suspiciously functional
Capping the drift at 3x3 for now -- it's a running tally. We do 9 checks
per tile (0,0 -- and then the 8 neighboring boxes), and update the drift
based on whichever resolved best.
There are surely some degenerate cases (e.g. maybe sometimes we'll miss
a series of tiles entirely and get "lost"), but this works better than
hoped in my tests.
It *probably* didn't work before just because I forgot (1, -1) and (-1,
1).