Maybe we shouldn't be trying to use edge hints for our deskew operation?
The homography does unpredictable things to our corner points --
probably because the logic to find the edges is sloppy. If we can make
the anchor scan more trustworthy/definitive, we should be able to rely
on the dewarp (which the edges can be used for? or not?) to handle the
rest.

This may also lead to getting rid of the edge hints altogether, if we
can use something else (e.g. the grid) to calculate lens distortion.