Not pretty, but the best I have right now...
The "tint_adjust" is -- unfortunately -- pretty hack-y, so it is
probably not really a good idea even though it kind of works.

The right answer would seem to be interative least squares on all color
values -- but the only way to get the latter right now is through
clustering (e.g. kmeans), and even then the results aren't great...