Experimenting with a sliding window for the color cutoff
There are currently 3 magic numbers:
1. the low cutoff for whether to count a pixel in a 3x3/6x6 towards our
average color for a tile. e.g. "75" (out of 255), where if no channel is
above that value, we treat the pixel as black and not contributing to
the color value of the tile.
2. the boundary (max) low value for color scaling. The idea is to treat
this as 0 in a sliding scale.
3. the boundary (min) high value for color scaling. The idea is to treat
this as 255 on a sliding scale.

(2) and (3) have been tightly coupled, with the
vague-srgb-but-not-really value of "48" as the prospective value for 0.
But all RGB channels are treated equally (and probably shouldn't),
scaling is done only with knowledge of the present tile (and maybe
shouldn't be), and the constant value of 48 seems overfit to my previous
testing.

So, let the games begin